Wrap client JS in DOMContentLoaded event listener
At frontendmentor.io we use puppeteer in order to analyise webpages for HTML
and accessibility issues, as well as to create screenshots. This particular
submission caused an unusual bug with puppeteer. After some investigation I
discovered that wrapping the client JS in a `DOMContentLoaded` event listener
resolved the issue.

I will raise this issue with puppeteer but it may be some time before I get a
response if at all. In the meantime this seems like a resonable solution.